requirements.txt
python-telegram-bot>=21.0.0
openai>=0.27.0